+++ LeMaitreCore/src/br/cefetmg/LeMaitre/model/service/TableManagement.java
@@ -14,9 +14,9 @@ import br.cefetmg.LeMaitre.model.exception.PersistenceException;
  * @author Thalesgsn
  */
 public interface TableManagement {
-    public Integer tableInsert(Table table) throws BusinessException, PersistenceException;
+    public Long tableInsert(Table table) throws BusinessException, PersistenceException;
     public boolean tableUpdate(Table table) throws BusinessException, PersistenceException;
-    public boolean tableRemove(Integer tableID) throws PersistenceException;
-    public Table getTableByID(Integer tableID) throws PersistenceException;
-    public boolean containsThisTableID(Integer tableID) throws PersistenceException;
+    public boolean tableRemove(Long tableID) throws PersistenceException;
+    public Table getTableByID(Long tableID) throws PersistenceException;
+    public boolean containsThisTableID(Long tableID) throws PersistenceException;
 }
